% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M4.R
\docType{data}
\name{M4}
\alias{M4}
\title{M4 Competiotion data}
\format{A list of  objects with\code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://www.m4.unic.ac.cy/the-dataset/}{M4 Competition Dataset}
}
\usage{
data(M4)
}
\description{
Featuring 100000 of diffrerent domains
}
\examples{
data(M4)
names(M4[[1]])
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.m4.unic.ac.cy/resources/makridakis-competitions/}{M4 Competition Web})
}
\keyword{datasets}
