% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M4.R
\docType{data}
\name{M4_norm}
\alias{M4_norm}
\title{Normalized M4 Dataset}
\format{
M4_norm is a list of 100000 series, where all numeric values have been min-max scaled.
Each series within \code{M4_norm} retains the same structure as \code{M4}, with scaled values:
\describe{
  \item{x}{A normalized time series of length \code{n} (the historical data, scaled between 0 and 1)}
  \item{xx}{A normalized time series of length \code{h} (the future data, scaled between 0 and 1)}
  \item{pt_ff}{A 25xh matrix of min-max scaled forecasts from the top 25 submissions in the M4 Competition.}
  \item{up_ff}{A min-max scaled matrix containing the upper bound of the prediction intervals.}
  \item{low_ff}{A min-max scaled matrix containing the lower bound of the prediction intervals.}
}
}
\usage{
data(M4_norm)
}
\description{
This dataset contains the normalized version of the M4 dataset.
If not found, it is automatically created using min-max scaling.
}
\note{
The normalization uses min-max scaling independently for each time series.
The original minimum and maximum values are stored to allow inverse transformation.
}
\examples{
data(M4_norm)
names(M4_norm[[1]])

}
\references{
(\href{https://www.m4.unic.ac.cy/resources/makridakis-competitions/}{M4 Competition Web})
(\href{https://github.com/M4Competition/M4-methods}{M4 Competition GitHub Repository})
}
\keyword{datasets}
