% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_max_scaler}
\alias{min_max_scaler}
\title{Min-Max Scaling Function}
\usage{
min_max_scaler(x, x_min, x_max)
}
\arguments{
\item{x}{A numeric vector or matrix to be scaled.}

\item{x_min}{The minimum value to use for scaling.}

\item{x_max}{The maximum value to use for scaling.}
}
\value{
A scaled numeric vector or matrix between 0 and 1.
}
\description{
Scales a numeric vector or matrix between 0 and 1 using provided x_min and x_max values.
}
