% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_database}
\alias{normalize_database}
\title{Normalize an entire dataset}
\usage{
normalize_database(dataset)
}
\arguments{
\item{dataset}{A list of time series entries.}
}
\value{
A dataset with all series entries normalized.
}
\description{
Applies `normalize_seriesentry()` to each series in a dataset.
}
