% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_M4_norm.R
\name{prepare_M4_norm}
\alias{prepare_M4_norm}
\title{Load or Create the Normalized M4 Dataset}
\usage{
prepare_M4_norm()
}
\value{
The `M4_norm` object, a normalized version of the `M4` dataset.
}
\description{
Ensures that `M4_norm` is available. If not, normalizes `M4`, saves it, and loads it.
}
